feat: EmulatorJS support - importing of BIOS files #15

+++ gaseous-server/Models/Signatures_Games.cs
@@ -236,7 +236,7 @@ namespace gaseous_server.Models
 
         public class SignatureFlags
         {
-            public int IGDBPlatformId { get; set; }
+            public long IGDBPlatformId { get; set; }
             public string IGDBPlatformName { get; set; }
         }
     }
